devops: Extract auth and env switcher into reusable components
