+++ app/layout.tsx
@@ -1,21 +1,10 @@
-import {
-  AppBar,
-  Button,
-  Menu,
-  MenuItem,
-  Toolbar,
-  Typography,
-} from '@mui/material';
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter';
 
 import type { Metadata } from 'next';
 
 import '@/globals.css';
+import AuthProvider from '@/_components/AuthProvider';
 import SystemThemeProvider from '@/_components/SystemThemeProvider';
-import {
-  TargetEnvironment,
-  TargetEnvironmentProvider,
-} from '@/_components/TargetEnvironment';
 
 export const metadata: Metadata = {
   title: 'Launchpad',
@@ -32,23 +21,9 @@ export default function RootLayout({
       <body>
         <AppRouterCacheProvider>
           <SystemThemeProvider>
-            <TargetEnvironmentProvider>
-              <AppBar>
-                <Toolbar>
-                  <Typography
-                    variant='h6'
-                    sx={{
-                      padding: '1rem'
-                    }}
-                  >
-                    🚀 this is how the world ends
-                  </Typography>
-
-                  <TargetEnvironment />
-                </Toolbar>
-              </AppBar>
+            <AuthProvider>
               {children}
-            </TargetEnvironmentProvider>
+            </AuthProvider>
           </SystemThemeProvider>
         </AppRouterCacheProvider>
       </body>
