+++ app/page.tsx
@@ -1,5 +1,4 @@
 'use client';
-import { signInWithPopup } from "firebase/auth";
 import Image from 'next/image';
 import {
   Button,
@@ -9,9 +8,7 @@ import {
   Stack,
   Typography,
 } from '@mui/material';
-import { useState } from 'react';
-
-import { auth, GoogleAuthProvider } from "#/firebase";
+import { useContext, useState } from 'react';
 
 import {
   Environment,
@@ -27,9 +24,10 @@ import {
   RevertButton,
 } from '@/_components/action-buttons';
 import ActiveDeployments from '@/_components/ActiveDeployments';
+import AppBar from '@/_components/AppBar';
+import { AUTH_CONTEXT } from '@/_components/AuthProvider';
 import EligiblePullRequests from '@/_components/EligiblePullRequests';
 import DeploymentHistory from '@/_components/DeploymentHistory';
-import useAuth from '@/_components/utils/useAuth';
 
 import styles from './page.module.css';
 
@@ -64,35 +62,34 @@ const Actions = {
 };
 
 export default function Home() {
-  var currentUser = useAuth({
-    onLogout: function login() {
-      signInWithPopup(auth, GoogleAuthProvider).then(() => console.log('signed in'));
-    },
-  });
-
+  var currentUser = useContext(AUTH_CONTEXT);
   if (!currentUser) {
     return (
-      <main className={styles.main}>
-        <CircularProgress />
-      </main>
+      <AppBar>
+        <main className={styles.main}>
+          <CircularProgress />
+        </main>
+      </AppBar>
     );
   }
 
   return (
-    <main className={styles.main}>
-      <Stack spacing={10}>
-        {/* NOTE(dabrady) Add component filter as we grow. */}
-        <EligiblePullRequests
-          components={DEPLOYABLE_COMPONENTS}
-          actions={Actions}
-        />
+    <AppBar withEnvSwitcher>
+      <main className={styles.main}>
+        <Stack spacing={10}>
+          {/* NOTE(dabrady) Add component filter as we grow. */}
+          <EligiblePullRequests
+            components={DEPLOYABLE_COMPONENTS}
+            actions={Actions}
+          />
 
-        <ActiveDeployments
-          components={DEPLOYABLE_COMPONENTS}
-        />
+          <ActiveDeployments
+            components={DEPLOYABLE_COMPONENTS}
+          />
 
-        <DeploymentHistory />
-      </Stack>
-    </main>
+          <DeploymentHistory />
+        </Stack>
+      </main>
+    </AppBar>
   );
 }
